requirements.txt
telethon==1.31.0
apscheduler==3.9.1
python-dotenv==1.0.0
